MDL-61743 core_privacy: Add context level checks for all users deletion

+++ mod/choice/classes/privacy/provider.php
@@ -184,6 +184,11 @@ class provider implements
         if (empty($context)) {
             return;
         }
+
+        if (!$context instanceof \context_module) {
+            return;
+        }
+
         $instanceid = $DB->get_field('course_modules', 'instance', ['id' => $context->instanceid], MUST_EXIST);
         $DB->delete_records('choice_answers', ['choiceid' => $instanceid]);
     }
@@ -202,6 +207,10 @@ class provider implements
 
         $userid = $contextlist->get_user()->id;
         foreach ($contextlist->get_contexts() as $context) {
+
+            if (!$context instanceof \context_module) {
+                return;
+            }
             $instanceid = $DB->get_field('course_modules', 'instance', ['id' => $context->instanceid], MUST_EXIST);
             $DB->delete_records('choice_answers', ['choiceid' => $instanceid, 'userid' => $userid]);
         }
